requirements.txt
fastapi
uvicorn
kiteconnect
pandas
pytz
python-dotenv
pyotp
requests
schedule